Boldog éj

+++ Assets/Scripts/Level/LevelObjectData.cs
@@ -1,7 +1,7 @@
 ﻿public class LevelObjectData {
 
 	public string uniqueID;
-	public int posX, posY, posZ;
-	public int rotX, rotY, rotZ;
+	public float posX, posY, posZ;
+	public float rotX, rotY, rotZ;
 
 }
